+++ src/core/ConfigurationResolver.ts
@@ -20,6 +20,12 @@ export interface Configuration {
   gamePath: string;
 }
 
+export type ConfigurationScope =
+  | vscode.Uri
+  | vscode.WorkspaceFolder
+  | vscode.TaskScope
+  | undefined;
+
 /**
  * ConfigurationResolver is responsible for resolving the final PlaydateSDK and
  * game configuration. It resolves configuration from the VS Code workspace
@@ -27,18 +33,21 @@ export interface Configuration {
  */
 @injectable()
 export class ConfigurationResolver {
-  async resolve(
-    scope: vscode.WorkspaceFolder | vscode.TaskScope | undefined
-  ): Promise<Configuration | undefined> {
+  async resolve(scope: ConfigurationScope): Promise<Configuration | undefined> {
     const workspaceFolder = getWorkspaceRoot(scope);
     if (!workspaceFolder) {
       return undefined;
     }
     const workspaceRoot = workspaceFolder.uri.fsPath;
 
-    let { sdkPath, sourcePath, outputPath, productName } =
-      vscode.workspace.getConfiguration("playdate-debug", workspaceFolder);
+    const {
+      sdkPath: sdkPathConfig,
+      sourcePath: sourcePathConfig,
+      outputPath: outputPathConfig,
+      productName: productNameConfig,
+    } = vscode.workspace.getConfiguration("playdate-debug", workspaceFolder);
 
+    let sdkPath = sdkPathConfig;
     if (!sdkPath) {
       sdkPath = await getSDKPath();
     } else {
@@ -47,19 +56,26 @@ export class ConfigurationResolver {
 
     const sdkVersion = await getSDKVersion(sdkPath);
 
+    let sourcePath = sourcePathConfig;
     if (!sourcePath) {
       sourcePath = path.resolve(workspaceRoot, "source");
     }
     sourcePath = toAbsolute(workspaceRoot, sourcePath);
 
+    let outputPath = outputPathConfig;
     if (!outputPath) {
       outputPath = workspaceRoot;
     }
     outputPath = toAbsolute(workspaceRoot, outputPath);
 
+    let productName = productNameConfig;
     if (!productName) {
-      const pdxInfo = await getPDXInfo(sourcePath);
-      productName = pdxInfo.name;
+      try {
+        const pdxInfo = await getPDXInfo(sourcePath);
+        productName = pdxInfo.name;
+      } catch (err) {
+        // noop
+      }
     }
 
     const productPath = path.resolve(outputPath, productName);