+++ src/debug/ProxyDebugAdapterDescriptorFactory.ts
@@ -3,7 +3,10 @@ import * as net from "net";
 import { inject, injectable } from "inversify";
 import * as vscode from "vscode";
 
-import { SIMULATOR_DEBUG_PORT } from "../constants";
+import {
+  CROSS_PLATFORM_DEBUG_SDK_VERSION,
+  SIMULATOR_DEBUG_PORT,
+} from "../constants";
 import { ConfigurationResolver } from "../core";
 import { WaitForDebugPortOptions } from "../util";
 
@@ -32,18 +35,18 @@ export class ProxyDebugAdapterDescriptorFactory
   ): Promise<vscode.DebugAdapterDescriptor | undefined> {
     this.server?.close();
 
+    const { disableWorkarounds, logDebugAdapter, retryTimeout, maxRetries } =
+      session.configuration;
+
     const config = await this.config.resolve(session.workspaceFolder);
     if (!config) {
       return undefined;
     }
 
     const { sdkVersion } = config;
-    if (sdkVersion >= "1.13.0") {
+    if (sdkVersion >= CROSS_PLATFORM_DEBUG_SDK_VERSION) {
       return new vscode.DebugAdapterServer(SIMULATOR_DEBUG_PORT);
     }
-
-    const { disableWorkarounds, logDebugAdapter, retryTimeout, maxRetries } =
-      session.configuration;
 
     const options: Partial<WaitForDebugPortOptions> = {};
     if (retryTimeout !== undefined) {