+++ src/pdc/index.ts
@@ -1,1 +1,2 @@
 export * from "./PDCModule";
+export * from "./PDCTaskProvider";