+++ src/simulator/index.ts
@@ -1,2 +1,3 @@
 export * from "./SimulatorModule";
 export * from "./getKillSimulatorCommand";
+export * from "./SimulatorTaskProvider";