+++ src/util/index.ts
@@ -1,5 +1,6 @@
 export * from "./exec";
 export * from "./quote";
+export * from "./runTask";
 export * from "./toAbsolute";
 export * from "./wait";
 export * from "./waitForDebugPort";